refactor(native): move runtime controller into subdomain folder
