+++ src/native/engine/runtime/tick/tick_post_phase.c
@@ -1,6 +1,6 @@
 #include "tick_post_phase.h"
 
-#include "../controller_sync.h"
+#include "../controller/controller_sync.h"
 
 void runtime_tick_post_phase_execute(World *world,
                                      ControllerInstance **controllers,
