+++ tests/native/runtime_controller_attach_test.c
@@ -1,4 +1,4 @@
-#include "runtime/controller_attach.h"
+#include "runtime/controller/controller_attach.h"
 
 #include <stdio.h>
 #include <stdlib.h>
