+++ tests/native/runtime_controller_world_metrics_test.c
@@ -1,5 +1,5 @@
 #include "ai/controller.h"
-#include "runtime/controller_runtime.h"
+#include "runtime/controller/controller_runtime.h"
 #include "runtime/world/world_metrics.h"
 #include "world/world.h"
